Refactor IngameComponent
