+++ agot-bg-game-server/src/common/EntireGame.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 import GameState, { SerializedGameState } from "./GameState";
 import LobbyGameState, {
   SerializedLobbyGameState,
